<jiraID>优化/修复/横展/新增: 修改内容: 影响范围: 自测情况:

+++ mcu_app/Bsw/Communication/Common/Dcm/UDS/Dcm_UDS0x2E.c
@@ -463,7 +463,7 @@ static FUNC(Std_ReturnType, DCM_CODE) Dcm_Uds0x2EServiceConditionCheck(
     /*get the required DID from request message*/
     if (E_OK == ret)
     {
-        RecDid = (uint16)((uint16)((uint16)pMsgContext->pReqData[1]) << 8u) | ((uint16)pMsgContext->pReqData[2]);
+        RecDid = (uint16)( ((uint16)pMsgContext->pReqData[1] << 8u) | ((uint16)pMsgContext->pReqData[2]) );
         /*Determine if the DID is configured*/
         ret = DspInternalUDS0x2E_DidCheck(OpStatus, RecDid, Dcm_0x2EType, ErrorCode);
         if (DCM_E_PENDING == ret)
